# docs/kaayko-paddling-api-swagger.yaml
openapi: 3.0.3
info:
  title: Kaayko Paddling Intelligence APIs
  version: "2.1.0"
  description: |
    # Kaayko Paddling Intelligence APIs v2.1.0
    
    **Enterprise-grade paddling condition intelligence powered by machine learning and real-time marine data integration**
    
    ---
    
    ## 🎯 Executive Summary
    
    The Kaayko Paddling Intelligence APIs provide ultra-fast, machine learning-powered paddling condition assessments with comprehensive safety analysis. Our platform processes real-time weather, marine, and environmental data to deliver actionable insights for paddlers across 18+ premium locations.
    
    ### Core Capabilities
    
    | API Endpoint | Purpose | Response Time | Key Features |
    |--------------|---------|---------------|--------------|
    | `paddlingOut` | Location Discovery | ~100ms | Comprehensive location database with amenities |
    | `paddleScore` | Real-time Conditions | ~250ms | Current conditions with ML-powered safety ratings |
    | `fastForecast` | 3-Day Predictions | <4s fresh / <1s cached | Hourly forecasts with penalty breakdown |
    | `forecast` | Premium Analytics | Variable | Advanced analytics for professional use |
    
    ### 🚀 Version 2.1.0 Enhancements
    
    - **🎯 Precision Rating System**: 0.5-increment ratings (1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0)
    - **🌊 Marine Data Integration**: Real-time wave height and water temperature analysis
...
    
    ## 🧠 Machine Learning Intelligence
    
...
    ```
...
    
...
servers:
...
security:
  - ApiKeyAuth: []
...
paths:
  /paddlingOut:
    get:
      tags:
        - paddlingOut
      operationId: getPaddlingLocations
      summary: Retrieve all paddling locations
      description: |
        **Comprehensive paddling location database with amenities and coordinates**
        
        This endpoint provides access to our curated database of premium paddling locations across North America. Each location includes precise coordinates, detailed amenities information, and multimedia resources.
        
        ### Key Features
        - ✅ **18+ Premium Locations** - Hand-selected paddling destinations
        - 🎯 **Precise Coordinates** - GPS coordinates for navigation systems
        - 🅿️ **Amenity Details** - Parking, restrooms, and facility information
        - 📸 **Rich Media** - Photos and instructional videos
        - 🏞️ **Location Context** - Difficulty ratings and paddling tips
        
        ### Use Case
        Use the coordinates from this endpoint for all subsequent API calls (`paddleScore`, `fastForecast`, `forecast`) to ensure marine data accuracy and optimal predictions.
        
        ### Response Time
        **~100ms** - Optimized for mobile and web applications
        
      responses:
        '200':
          description: Successfully retrieved paddling locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaddlingLocation'
              examples:
                complete_response:
                  summary: Complete locations response
                  description: Array of paddling locations with full details
                  value:
                    - id: "trinity_river"
                      lakeName: "Trinity River"
                      title: "Trinity River"
                      subtitle: "Irving, Texas"
                      text: "Solitude, birds and a good workout with consistent flow"
                      location:
                        latitude: 32.881187
                        longitude: -96.929937
                      parkingAvl: "Y"
                      restroomsAvl: "Y"
                      youtubeURL: "https://youtube.com/watch?v=trinity_river_guide"
                      imgSrc: 
                        - "https://firebasestorage.googleapis.com/v0/b/kaaykostore.appspot.com/o/images%2Fpaddling_out%2Ftrinity_river_1.jpg?alt=media"
                        - "https://firebasestorage.googleapis.com/v0/b/kaaykostore.appspot.com/o/images%2Fpaddling_out%2Ftrinity_river_2.jpg?alt=media"
                    - id: "cottonwood_lake"
                      lakeName: "Cottonwood Lake"
                      title: "Cottonwood Lake"
                      subtitle: "Buena Vista, Colorado"
                      text: "High altitude alpine paddling with stunning mountain views"
                      location:
                        latitude: 38.781063
                        longitude: -106.277812
                      parkingAvl: "Y"
                      restroomsAvl: "N"
                      youtubeURL: ""
                      imgSrc: 
                        - "https://firebasestorage.googleapis.com/v0/b/kaaykostore.appspot.com/o/images%2Fpaddling_out%2Fcottonwood_lake_1.jpg?alt=media"
                    - id: "lake_crescent"
                      lakeName: "Lake Crescent"
                      title: "Lake Crescent"
                      subtitle: "Olympic National Park, Washington"
                      text: "Crystal clear glacial lake surrounded by old-growth forest"
                      location:
                        latitude: 48.052813
                        longitude: -123.870438
                      parkingAvl: "Y"
                      restroomsAvl: "Y"
                      youtubeURL: ""
                      imgSrc:
                        - "https://firebasestorage.googleapis.com/v0/b/kaaykostore.appspot.com/o/images%2Fpaddling_out%2Flake_crescent_1.jpg?alt=media"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Server error"
                details: "Database connection failed"
                timestamp: "2025-08-23T10:30:00Z"
...
  /forecast:
    get:
      tags:
        - forecast
      operationId: getForecast
      summary: Generate comprehensive weather forecast with ML predictions (Premium)
      description: |
        **🔒 PREMIUM FORECAST API - Advanced comprehensive weather analysis with ML intelligence**
        
        This premium endpoint generates comprehensive weather forecasts with ML predictions, used by scheduled jobs and premium users. Results are automatically cached for the `fastForecast` API to serve to frontend applications.
        
        ### � Advanced Features
        - **Comprehensive Analysis**: Detailed weather data with ML-powered paddle predictions
        - **Auto-Caching**: Results cached for 2 hours to optimize `fastForecast` performance
        - **Safety Analysis**: Professional safety level assessments with color-coded warnings
        - **Batch Processing**: Supports scheduled background processing for multiple locations
        - **Premium Access**: Rate-limited for internal and premium users only
        
        ### 🎯 Use Cases
        - **Scheduled Jobs**: Background forecast generation for popular paddling spots
        - **Premium Users**: Advanced forecast analysis with detailed ML insights
        - **Data Pipeline**: Source data for high-performance `fastForecast` endpoint
        - **Research**: Comprehensive weather-to-paddle correlations
        
        ### ⚡ Performance Characteristics
        - **Rate Limit**: 10 requests per minute (premium tier)
        - **Processing**: 3-5 seconds for fresh data generation
        - **Caching**: 2-hour cache lifetime in Firestore
        - **ML Integration**: Full ML service pipeline with penalty analysis
        
        ### 🔄 Relationship to fastForecast
        This endpoint generates the comprehensive data that powers the ultra-fast `fastForecast` API. Use `forecast` for data generation and analysis, use `fastForecast` for high-performance frontend delivery.
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: double
            minimum: -90.0
            maximum: 90.0
            example: 32.881187
          description: |
            **Latitude coordinate in decimal degrees**
            
            Must be a valid latitude between -90.0 and 90.0.
            
            **Example**: `32.881187` (Trinity River, Texas)
        - name: lng
          in: query
          required: true
          schema:
            type: number
            format: double
            minimum: -180.0
            maximum: 180.0
            example: -96.929937
          description: |
            **Longitude coordinate in decimal degrees**
            
            Must be a valid longitude between -180.0 and 180.0.
            
            **Example**: `-96.929937` (Trinity River, Texas)
      responses:
        '200':
          description: Successfully retrieved premium forecast analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForecastResponse'
              examples:
                premium_analytics:
                  summary: Premium Analytics Response - Colorado High Altitude
                  description: Comprehensive analytics with extended data and marine integration
                  value:
                    success: true
                    data:
                      location: "38.781063,-106.277812"
                      locationDetails:
                        name: "Cottonwood Lake"
                        region: "Buena Vista, Colorado"
                        elevation: 10040
                        timezone: "America/Denver"
                        coordinates:
                          latitude: 38.781063
                          longitude: -106.277812
                      current:
                        temperature: 68.2
                        windSpeed: 6.8
                        conditions: "Partly Cloudy"
                        marine:
                          waveHeight: null
                          waterTemp: null
                          available: false
                        paddleAnalysis:
                          rating: 4.5
                          interpretation: "Excellent conditions with minor considerations"
                          confidence: "high"
                          factors: ["Optimal temperature", "Light winds", "Good visibility"]
                      forecast:
                        - date: "2025-08-23"
                          summary:
                            maxTemp: 75.8
                            minTemp: 52.1
                            avgWindSpeed: 6.2
                            precipitation: 0.0
                            paddleWindow: "Excellent all day"
                            bestRating: 5.0
                            avgRating: 4.6
                            conditions: "Mostly sunny"
                          hourly:
                            - hour: "08:00"
                              temperature: 62.1
                              windSpeed: 4.2
                              windDirection: "SW"
                              humidity: 35
                              cloudCover: 15
                              uvIndex: 2.8
                              visibility: 25
                              beaufortScale: 2
                              prediction:
                                rating: 4.5
                                originalRating: 4.5
                                penaltiesApplied: []
                                totalPenalty: 0.0
                                mlModelUsed: true
                      analytics:
                        historicalComparison:
                          avgRatingThisTimeLastYear: 4.2
                          temperatureTrend: "2.3°F above average"
                          windTrend: "15% calmer than typical"
                        seasonalContext:
                          season: "Late Summer"
                          optimalMonths: ["June", "July", "August", "September"]
                          currentRanking: "Excellent (top 15%)"
                        riskAssessment:
                          overallRisk: "Low"
                          primaryConcerns: []
                          recommendations: ["Ideal conditions for all skill levels"]
                    metadata:
                      generated: "2025-08-23T15:45:22.120Z"
                      cached_until: "2025-08-23T16:45:22.120Z"
                      source: "premium-forecast-v2.1"
                      processingTimeMs: 2847
                      dataFreshness: "live"
                      modelVersion: "enterprise-ml-v2.1.0"
                      billingTier: "premium"
        '401':
          description: Authentication required or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_api_key:
                  summary: Missing API Key
                  value:
                    success: false
                    error: "Authentication required"
                    code: "MISSING_API_KEY"
                    details: "X-API-Key header is required for premium endpoints"
                    suggestion: "Contact api-support@kaayko.com for enterprise access"
                    timestamp: "2025-08-23T15:45:22.120Z"
                invalid_api_key:
                  summary: Invalid API Key
                  value:
                    success: false
                    error: "Invalid authentication"
                    code: "INVALID_API_KEY"
                    details: "The provided API key is invalid or expired"
                    suggestion: "Verify your API key or contact support for renewal"
                    timestamp: "2025-08-23T15:45:22.120Z"
        '403':
          description: Access forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Access forbidden"
                code: "INSUFFICIENT_PERMISSIONS"
                details: "Your account does not have access to premium endpoints"
                suggestion: "Upgrade to enterprise plan for premium API access"
                timestamp: "2025-08-23T15:45:22.120Z"
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Rate limit exceeded"
                code: "PREMIUM_RATE_LIMIT_EXCEEDED"
                details: "Premium tier allows 100 requests per hour"
                retryAfter: 1247
                timestamp: "2025-08-23T15:45:22.120Z"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        **Enterprise API Key Authentication**
        
        Contact api-support@kaayko.com for enterprise access and API key provisioning.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        **OAuth 2.0 Bearer Token Authentication**
        
        For service-to-service authentication and enterprise integrations.
...
  schemas:
    PaddlingLocation:
      type: object
      title: Paddling Location Details
      description: Complete information about a paddling location including coordinates, amenities, and media
      required:
        - id
        - title
        - location
      properties:
        id:
          type: string
          description: Unique location identifier for API references
          pattern: '^[a-z0-9_]+$'
          example: "trinity_river"
        title:
          type: string
          description: Human-readable location name
          minLength: 3
          maxLength: 100
          example: "Trinity River"
        subtitle:
          type: string
          description: Location context and geographical details
          maxLength: 200
          example: "Irving, Texas"
        text:
          type: string
          description: Paddling description, tips, and local insights
          maxLength: 500
          example: "Solitude, birds and a good workout with consistent flow year-round"
        location:
          type: object
          title: GPS Coordinates
          description: Precise GPS coordinates for navigation and API calls
          required:
            - latitude
            - longitude
          properties:
            latitude:
              type: number
              format: double
              minimum: -90.0
              maximum: 90.0
              description: Latitude in decimal degrees (WGS84)
              example: 32.881187
            longitude:
              type: number
              format: double  
              minimum: -180.0
              maximum: 180.0
              description: Longitude in decimal degrees (WGS84)
              example: -96.929937
        difficulty:
          type: string
          description: Paddling difficulty assessment
          enum: ["Easy", "Easy to Moderate", "Moderate", "Moderate to Advanced", "Advanced", "Expert"]
          example: "Easy to Moderate"
        elevation:
          type: integer
          description: Elevation above sea level in feet
          minimum: -100
          maximum: 15000
          example: 450
        features:
          type: array
          description: Key location features and characteristics
          items:
            type: string
          maxItems: 10
          example: ["Urban wildlife", "Year-round access", "Calm waters", "Scenic views"]
        parkingAvl:
          type: boolean
          description: Parking facilities available
          example: true
        restroomsAvl:
          type: boolean
          description: Restroom facilities available
          example: true
        permits_required:
          type: boolean
          description: Whether permits or fees are required
          default: false
          example: false
        youtubeURL:
          type: string
          format: uri
          description: Instructional or promotional YouTube video URL
          pattern: '^https://(www\.)?youtube\.com/watch\?v=.+$'
          example: "https://youtube.com/watch?v=trinity_river_guide"
        imgSrc:
          type: array
          description: High-resolution location images
          items:
            type: string
            format: uri
            description: Direct URL to location image
          maxItems: 5
          example: 
            - "https://kaayko.com/images/trinity_river_aerial.jpg"
            - "https://kaayko.com/images/trinity_river_launch.jpg"
        amenities:
          type: object
          description: Additional amenity details
          properties:
            picnicTables:
              type: boolean
              example: true
            grillsAvailable:
              type: boolean
              example: false
            shadedAreas:
              type: boolean
              example: true
            rentalEquipment:
              type: boolean
              example: false
...
    ErrorResponse:
      type: object
      title: Error Response
      description: Standardized error response format with detailed debugging information
      required:
        - success
        - error
        - code
        - timestamp
      properties:
        success:
          type: boolean
          description: Request success indicator (always false for errors)
          example: false
        error:
          type: string
          description: Human-readable error message
          maxLength: 200
          example: "Invalid coordinates"
        code:
          type: string
          description: Machine-readable error code for programmatic handling
          pattern: '^[A-Z_]+$'
          example: "INVALID_COORDINATES"
        details:
          type: string
          description: Additional technical details about the error
          maxLength: 500
          example: "Latitude must be between -90.0 and 90.0, longitude must be between -180.0 and 180.0"
        suggestion:
          type: string
          description: Helpful suggestion for resolving the error
          maxLength: 300
          example: "Use coordinates from paddlingOut API for accurate marine data"
        timestamp:
          type: string
          format: date-time
          description: UTC timestamp when error occurred
          example: "2025-08-23T15:45:22.120Z"
        requestId:
          type: string
          description: Unique identifier for this request (for support purposes)
          pattern: '^[a-f0-9-]+$'
          example: "req_8f9e2d1a-4b5c-6789-01ef-234567890abc"
        retryAfter:
          type: integer
          minimum: 0
          description: Seconds to wait before retrying (for rate limit errors)
          example: 60
        supportUrl:
          type: string
          format: uri
          description: URL for additional support or documentation
          example: "https://kaayko.com/docs/api/errors#invalid-coordinates"
...